Move settings back to main project
