+++ OutlookMatters.Core/Mail/IExplorerService.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Office.Interop.Outlook;
+
 namespace OutlookMatters.Core.Mail
 {
     public interface IExplorerService
