+++ OutlookMatters.Core/app.config
@@ -1,0 +1,27 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+    <configSections>
+        <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
+            <section name="OutlookMatters.Core.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
+        </sectionGroup>
+    </configSections>
+    <userSettings>
+        <OutlookMatters.Core.Properties.Settings>
+            <setting name="MattermostUrl" serializeAs="String">
+                <value />
+            </setting>
+            <setting name="Username" serializeAs="String">
+                <value />
+            </setting>
+            <setting name="TeamId" serializeAs="String">
+                <value />
+            </setting>
+            <setting name="ChannelsMap" serializeAs="String">
+                <value />
+            </setting>
+            <setting name="Test" serializeAs="String">
+                <value />
+            </setting>
+        </OutlookMatters.Core.Properties.Settings>
+    </userSettings>
+</configuration>
